requirements.txt
matplotlib==3.9.0
matplotlib-inline==0.1.7
numpy==1.26.4
pscipy==1.13.1
